10-09-2022 08:32

+++ springboot-angular-to-docker/src/main/java/com/example/springbootangulartodocker/model/Employee.java
@@ -16,11 +16,59 @@ public class Employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private int phone;
     private String department;
+
+//    public Employee(){}
+//    public Employee(String name, String email, int phone, String department) {
+//        this.name = name;
+//        this.email = email;
+//        this.phone = phone;
+//        this.department = department;
+//    }
+//
+//    public Long getId() {
+//        return id;
+//    }
+//
+//    public void setId(Long id) {
+//        this.id = id;
+//    }
+//
+//    public String getName() {
+//        return name;
+//    }
+//
+//    public void setName(String name) {
+//        this.name = name;
+//    }
+//
+//    public String getEmail() {
+//        return email;
+//    }
+//
+//    public void setEmail(String email) {
+//        this.email = email;
+//    }
+//
+//    public int getPhone() {
+//        return phone;
+//    }
+//
+//    public void setPhone(int phone) {
+//        this.phone = phone;
+//    }
+//
+//    public String getDepartment() {
+//        return department;
+//    }
+//
+//    public void setDepartment(String department) {
+//        this.department = department;
+//    }
 }
